requirement.txt
annotated-types==0.7.0
anyio==4.5.2
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.11
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
lxml==5.3.1
openai==1.66.3
pydantic==2.10.6
pydantic_core==2.27.2
PyPDF2==3.0.1
python-docx==1.1.2
python-multipart==0.0.20
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.44.0
tqdm==4.67.1
typing_extensions==4.12.2
uvicorn==0.33.0
